More optimization on the lowpass filter

+++ GlobalFilterTables.h
@@ -2,6 +2,5 @@
 #define FILTER_TABLES
 
 extern const uint16_t LPF_P_TABLE[];
-extern const float LPF_T4_TABLE[];
-
+extern const uint16_t LPF_T4_TABLE[];
 #endif
